Move functions for obtaining batch IDs
